+++ Mobile/starter_project_dartsmiths/lib/features/feed/home/domain/entity/home.dart
