+++ Mobile/starter_project_dartsmiths/lib/features/feed/home/domain/repository/home_repository.dart
@@ -1,5 +1,5 @@
 import 'package:dartsmiths/core/error/failures.dart';
-import 'package:dartsmiths/features/home/domain/entity/home.dart';
+import 'package:dartsmiths/features/feed/home/domain/entity/home.dart';
 import 'package:dartz/dartz.dart';
 
 abstract class HomeRepository {
