+++ Mobile/starter_project_dartsmiths/lib/features/feed/home/domain/usecase/home_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:dartsmiths/core/error/failures.dart';
 import 'package:dartsmiths/core/usecases/usecase.dart';
-import 'package:dartsmiths/features/home/domain/repository/home_repository.dart';
+import 'package:dartsmiths/features/feed/home/domain/repository/home_repository.dart';
 import 'package:dartz/dartz.dart';
 import '../entity/home.dart';
 class GetBytag extends UseCase<Home, Params> {
