+++ Mobile/starter_project_dartsmiths/test/features/feed/home/domain/model/home_model_test.dart
@@ -1,9 +1,9 @@
 import 'dart:convert';
 
-import 'package:dartsmiths/features/home/data/model/home.dart';
+import 'package:dartsmiths/features/feed/home/data/model/home.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../../core/utils/fixture_reader.dart';
+import '../../../../../core/utils/fixture_reader.dart';
 
 void main() {
   const homeModel = HomeModel(
